+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # GoodData Ruby SDK Changelog
 
 ## 0.6.21
+- Connect using SSO - GoodData#connect_sso
 
 ## 0.6.20
 - added iterators for retrieval of project, domain, and group objects